+++ app/Providers/Filament/AdminPanelProvider.php
@@ -38,6 +38,7 @@ class AdminPanelProvider extends PanelProvider
                 'PayPal',
                 'Events & Kunden',
                 'Transaktionen',
+                'Berichte',
                 'Exporte',
                 'Einstellungen',
             ])
@@ -49,6 +50,7 @@ class AdminPanelProvider extends PanelProvider
             ->widgets([
                 Widgets\AccountWidget::class,
                 \App\Filament\Widgets\DashboardStatsOverview::class,
+                \App\Filament\Widgets\SyncHealthWidget::class,
                 \App\Filament\Widgets\RevenueByDayChart::class,
                 \App\Filament\Widgets\RecentSyncRunsWidget::class,
             ])